pyproject.toml
[project]
name = "route-tracker"
version = "1.0.0"
description = "Live route tracking PWA for travelers"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.41",
    "pillow>=10.0.0",
]
